Move WebGPU allocator tests to dedicated file
